+++ src/foundry/client/core/index.d.mts
@@ -1,2 +1,1 @@
-import "./hooks.d.mts";
 import "./packages.d.mts";
