+++ src/foundry/client/hooks.d.mts
@@ -10,8 +10,9 @@ import type {
   DatabaseDeleteOperation,
   DatabaseUpdateOperation,
 } from "#common/abstract/_types.d.mts";
-import type CompendiumArt from "#client/helpers/media/compendium-art.mjs";
+import type CompendiumArt from "#client/helpers/media/compendium-art.d.mts";
 import type { Hooks as HookConfigs } from "#configuration";
+import type Hooks from "./helpers/hooks.d.mts";
 
 import AVSettings = foundry.av.AVSettings;
 import Application = foundry.appv1.api.Application;
